Set up early basic structure
First major commit of my portion of project.
Added 3 files:
-ClinicalTrialAPIUtil.js
-MainSearchQueryParams.js
-ClinicalTrialSearchScreen.js

The screen file is screen main search will take place on. Still mostly template for now with some buttons for testing and a loading indicator when API call is running.

Params file has strings for all the parameters in the API call in case they need to be referenced somewhere else. Also has an array of all fields being included in the result returned.

The util file is the class with static methods for making the API call(s)